Redirect clicks on login button icons to the login button itself
Visually, the icon positioned on the login button is part of the login
button, so clicking on the icon should be the same as clicking on the
button. As the icon is a different element this commit adds a
redirection of the clicks on the icon to the login button using
JavaScript.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>